TableViewController created to clean AdvancedJTable_NetworkElement and separate functionallity
